up to date

+++ src/main/java/io/github/Oleiva/dao/ListOfCustomersDao.java
@@ -3,7 +3,7 @@ package io.github.Oleiva.dao;
 import io.github.Oleiva.entity.CustomersEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ListOfCustomersDao  extends JpaRepository<CustomersEntity, Integer> {
+public interface ListOfCustomersDao  extends JpaRepository<CustomersEntity, Long> {
 
 //    List<SubscriptionModel> findByIssueId(int issueId);
 //
